Move hasClioDependencies to packageConfig
