+++ package/packageConfig.js
@@ -59,10 +59,27 @@ function getPackageDependencies() {
   return config.dependencies;
 }
 
+/**
+ * @method hasClioDependencies
+ * @returns {bool}
+ * @description Returns true if the project has at least one dependency listed
+ *              in the Package.json file.
+ */
+
+function hasClioDependencies() {
+  const dependencies = getPackageDependencies();
+  return (
+    !!dependencies &&
+    !!Object.keys(dependencies) &&
+    !!Object.keys(dependencies).length
+  );
+}
+
 module.exports = {
   getPackageConfig,
   writePackageConfig: writePackageConfig,
   addDependency,
   getPackageDependencies,
-  configFileName
+  configFileName,
+  hasClioDependencies
 };
